pyproject.toml
[project]
name = "specialk"
version = "0.1.0"
description = "text based style transfer with sequence models"
authors = [{ name = "@thien", email = "thien@users.noreply.github.com" }]
dependencies = [
    "absl-py==0.7.0",
    "accelerate==0.34.0",
    "aiohttp==3.9.5",
    "aiosignal==1.3.1",
    "annotated-types==0.7.0",
    "antlr4-python3-runtime==4.8",
    "appnope==0.1.4",
    "asttokens==2.4.1",
    "async-timeout==4.0.3",
    "attrs==23.2.0",
    "bayesian-optimization==1.0.1",
    "beautifulsoup4==4.7.1",
    "better-exceptions==0.3.3",
    "bitarray==2.9.2",
    "blis==0.7.11",
    "build==1.2.1",
    "cachecontrol==0.14.0",
    "catalogue==2.0.10",
    "certifi==2024.6.2",
    "cffi==1.16.0",
    "charset-normalizer==3.3.2",
    "cleo==2.1.0",
    "click==8.1.7",
    "cloudpathlib==0.18.1",
    "colorama==0.4.6",
    "comm==0.2.2",
    "confection==0.1.5",
    "contourpy==1.2.1",
    "crashtest==0.4.1",
    "cycler==0.12.1",
    "cymem==2.0.8",
    "cython==3.0.10",
    "datasets==2.20.0",
    "de-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/de_core_news_sm-3.7.0/de_core_news_sm-3.7.0-py3-none-any.whl",
    "debugpy==1.8.1",
    "decorator==5.1.1",
    "dill==0.3.8",
    "distlib==0.3.8",
    "docopt==0.6.2",
    "dulwich==0.21.7",
    "einops==0.8.0",
    "en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl",
    "evaluate==0.4.2",
    "exceptiongroup==1.2.1",
    "executing==2.0.1",
    "fairseq==0.12.2",
    "fastjsonschema==2.20.0",
    "fasttext @ git+https://github.com/facebookresearch/fastText.git@1142dc4c4ecbc19cc16eee5cdd28472e689267e6",
    "filelock==3.13.1",
    "fonttools==4.53.1",
    "frozenlist==1.4.1",
    "fsspec==2024.2.0",
    "ftfy==6.2.3",
    "gensim==4.3.2",
    "huggingface-hub==0.23.4",
    "hydra-core==1.0.7",
    "idna==3.7",
    "importlib-metadata==7.2.1",
    "importlib-resources==6.4.4",
    "iniconfig==2.0.0",
    "installer==0.7.0",
    "ipykernel==6.29.4",
    "ipynb==0.5.1",
    "ipython==8.18.1",
    "ipywidgets==8.1.3",
    "jaraco-classes==3.4.0",
    "jaxtyping==0.2.31",
    "jedi==0.19.1",
    "jinja2==3.1.3",
    "joblib==1.4.2",
    "jupyter-client==8.6.2",
    "jupyter-core==5.7.2",
    "jupyterlab-widgets==3.0.11",
    "keyring==24.3.1",
    "kiwisolver==1.4.5",
    "langcodes==3.4.0",
    "language-data==1.2.0",
    "lightning==2.3.1",
    "lightning-utilities==0.11.3.post0",
    "lxml==5.2.2",
    "marisa-trie==1.2.0",
    "markdown-it-py==3.0.0",
    "markupsafe==2.1.5",
    "matplotlib==3.9.2",
    "matplotlib-inline==0.1.7",
    "mdurl==0.1.2",
    "memory-profiler==0.61.0",
    "more-itertools==10.3.0",
    "mpmath==1.3.0",
    "msgpack==1.0.8",
    "multidict==6.0.5",
    "multiprocess==0.70.16",
    "murmurhash==1.0.10",
    "nest-asyncio==1.6.0",
    "networkx==3.2.1",
    "nltk==3.8.1",
    "numpy==1.26.4",
    "omegaconf==2.0.6",
    "packaging==24.1",
    "pandarallel==1.6.5",
    "pandas==2.2.2",
    "parso==0.8.4",
    "peft==0.12.0",
    "pexpect==4.9.0",
    "pillow==10.2.0",
    "pip==21.2.3",
    "pkginfo==1.11.1",
    "platformdirs==4.2.2",
    "pluggy==1.5.0",
    "poetry==1.8.3",
    "poetry-core==1.9.0",
    "poetry-plugin-export==1.8.0",
    "polars==1.4.1",
    "portalocker==2.8.2",
    "pot==0.9.3",
    "preshed==3.0.9",
    "prompt-toolkit==3.0.47",
    "protobuf==5.27.1",
    "psutil==6.0.0",
    "ptyprocess==0.7.0",
    "pure-eval==0.2.2",
    "pyarrow==16.1.0",
    "pyarrow-hotfix==0.6",
    "pybind11==2.13.3",
    "pycparser==2.22",
    "pydantic==2.8.2",
    "pydantic-core==2.20.1",
    "pyemd==1.0.0",
    "pyfakefs==5.6.0",
    "pygments==2.18.0",
    "pyparsing==3.1.2",
    "pyproject-hooks==1.1.0",
    "pytelegrambotapi==4.20.0",
    "pytest==8.2.2",
    "pytest-watch==4.2.0",
    "python-dateutil==2.9.0.post0",
    "pytorch-lightning==2.3.1",
    "pytz==2024.1",
    "pyyaml==6.0.1",
    "pyzmq==26.0.3",
    "rapidfuzz==3.9.3",
    "regex==2024.5.15",
    "requests==2.32.3",
    "requests-toolbelt==1.0.0",
    "rich==13.7.1",
    "rouge-score==0.1.2",
    "sacrebleu==2.4.2",
    "sacremoses==0.1.1",
    "safetensors==0.4.3",
    "schedulefree==1.2.7",
    "scikit-learn==1.5.0",
    "scipy==1.12.0",
    "sentencepiece==0.2.0",
    "setuptools==70.1.1",
    "shellingham==1.5.4",
    "six==1.16.0",
    "smart-open==7.0.4",
    "soupsieve==2.5",
    "spacy==3.7.5",
    "spacy-legacy==3.0.12",
    "spacy-loggers==1.0.5",
    "srsly==2.4.8",
    "stack-data==0.6.3",
    "structlog==24.2.0",
    "sympy==1.12",
    "tabulate==0.9.0",
    "telebot==0.0.5",
    "tensorboardx==2.6.2.2",
    "thinc==8.2.5",
    "threadpoolctl==3.5.0",
    "tokenizers==0.19.1",
    "tomli==2.0.1",
    "tomlkit==0.12.5",
    "toolz==0.12.1",
    "torch==2.3.1",
    "torchaudio==2.3.1",
    "torchdata==0.7.1",
    "torchmetrics==1.4.0.post0",
    "torchtext==0.18.0",
    "torchvision==0.18.1",
    "tornado==6.4.1",
    "tqdm==4.66.4",
    "traitlets==5.14.3",
    "transformers==4.43.3",
    "trove-classifiers==2024.5.22",
    "typeguard==2.13.3",
    "typer==0.12.3",
    "typing-extensions==4.12.2",
    "tzdata==2024.1",
    "urllib3==2.2.1",
    "virtualenv==20.26.3",
    "wasabi==1.1.3",
    "watchdog==4.0.1",
    "wcwidth==0.2.13",
    "weasel==0.4.1",
    "widgetsnbextension==4.0.11",
    "wrapt==1.16.0",
    "xattr==1.1.0",
    "xxhash==3.4.1",
    "yarl==1.9.4",
    "zipp==3.19.2",
    "pytest-xdist>=3.6.1",
]
readme = "readme.md"
requires-python = ">= 3.9"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
universal = true
virtual = true
dev-dependencies = []

[tool.rye.scripts]
# These three options are equivalent:
download-deps = 'python -m nltk.downloader cmudict averaged_perceptron_tagger'
tests-light = 'pytest -v -m "not heavyweight"'
tests-heavy = 'pytest -v -m "not lightweight" -n auto'
tests = 'pytest -v -n auto'

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["specialk"]

[tool.ruff.lint]
ignore=["F722"]

[tool.pytest.ini_options]
testpaths = "./tests"
pythonpath = "."
addopts = "-ra -q"
filterwarnings = ["ignore:.*U.*mode is deprecated:DeprecationWarning", "ignore::DeprecationWarning"]
markers = ["lightweight: marks tests as lightweight (deselect with '-m \"not lightweight\"')","heavyweight: marks tests that require large model downloads"]

[tool.coverage.run]
source = ["specialk"]
omit = ["tests/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if __name__ == .__main__.:",
    "raise NotImplementedError",
    "pass",
    "except ImportError:"]

